pyproject.toml
[tool.poetry]
name = "autokm"
version = "0.1.0"
description = ""
authors = ["Guilherme <gui.dias.alves20@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
django-cors-headers = "^4.4.0"
django = "^5.1"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
